refactor: moving events package into process package
