+++ src/scripts/process/events/detectPageLoad.tab.js
