+++ src/scripts/process/events/exportAll.bg.js
@@ -1,6 +1,6 @@
-import {clickOnListElt} from "../units/interface/interact-DOM/interact.tab";
+import {clickOnListElt} from "../../units/interface/interact-DOM/interact.tab";
 
-import {launchScrapping} from "../process/scraper/launchScrapping.tab";
+import {launchScrapping} from "../scraper/launchScrapping.tab";
 
 export function exportAllThreadsListener() {
   let currentIndex = 0;
