+++ src/scripts/process/events/notify.bg.js
