+++ src/scripts/process/events/uiEnhancer.tab.js
@@ -1,6 +1,6 @@
-import {appendModalContainer} from "../units/interface/modals/appendModalContainer.tab";
-import ModalUpdate from "../units/interface/modals/ModalUpdate.tab";
-import ModalMessage from "../units/interface/modals/ModalMessage.tab";
+import {appendModalContainer} from "../../units/interface/modals/appendModalContainer.tab";
+import ModalUpdate from "../../units/interface/modals/ModalUpdate.tab";
+import ModalMessage from "../../units/interface/modals/ModalMessage.tab";
 
 
 export async function uiEnhancer(domain) {
