+++ src/scripts/process/events/updateButtons.tab.js
@@ -1,3 +1,6 @@
+import {setBtnsExport} from "../../units/interface/uiEnhancer/phind/exportButtons/styleExportButtons";
+import {btnBarAllInline} from "../../units/interface/uiEnhancer/phind/exportButtons/changeButtons";
+
 export function updateButtonsOnResize() {
   window.addEventListener('resize', function () {
     setBtnsExport(isExporting, exportAllThreadsSideBtn, exportAllThreadsTopBtn, stopExportAllThreadsSideBtn, stopExportAllThreadsTopBtn);
